+++ src/rer/ufficiostampa/restapi/services/search_parameters/get.py
@@ -10,20 +10,40 @@ from zope.publisher.interfaces import IPublishTraverse
 from zope.schema.interfaces import IVocabularyFactory
 
 
-def getVocabularyTermsForForm(vocab_name, context):
+def getVocabularyTermsForForm(vocab_name):
     """
     Return the values of vocabulary
     """
+    portal = api.portal.get()
     utility = getUtility(IVocabularyFactory, vocab_name)
 
     values = []
 
-    vocab = utility(context)
+    vocab = utility(portal)
 
     for entry in vocab:
         if entry.title != u"select_label":
             values.append({"value": entry.value, "label": entry.title})
+    values[0]["isFixed"] = True
     return values
+
+
+def getArguments():
+    legislatures = getVocabularyTermsForForm(
+        vocab_name="rer.ufficiostampa.vocabularies.legislatures",
+    )
+
+    res = {}
+    for legislature in legislatures:
+        key = legislature.get("value", "")
+        arguments = []
+        for brain in api.content.find(legislature=key):
+            for argument in brain.arguments:
+                argument_dict = {"value": argument, "label": argument}
+                if argument_dict not in arguments:
+                    arguments.append(argument_dict)
+        res[key] = sorted(arguments, key=lambda x: x["label"])
+    return res
 
 
 def getTypesValues():
@@ -43,8 +63,9 @@ def getTypesDefault():
 
 def getSearchFields():
     request = getRequest()
-    portal = api.portal.get()
-
+    legislatures = getVocabularyTermsForForm(
+        vocab_name="rer.ufficiostampa.vocabularies.legislatures",
+    )
     return [
         {
             "id": "SearchableText",
@@ -58,7 +79,8 @@ def getSearchFields():
         {
             "id": "portal_type",
             "label": translate(
-                _("label_portal_type", default="Type"), context=request,
+                _("label_portal_type", default="Type"),
+                context=request,
             ),
             "help": "",
             "type": "checkbox",
@@ -78,29 +100,29 @@ def getSearchFields():
         {
             "id": "legislature",
             "label": translate(
-                _("label_legislature", default="Legislature"), context=request,
-            ),
-            "help": "",
-            "type": "select",
-            "multivalued": True,
-            "options": getVocabularyTermsForForm(
-                context=portal,
-                vocab_name="rer.ufficiostampa.vocabularies.legislatures",
-            ),
-        },
-        {
-            "id": "arguments",
-            "label": translate(
-                _("legislature_arguments_label", default="Arguments"),
+                _("label_legislature", default="Legislature"),
                 context=request,
             ),
             "help": "",
             "type": "select",
             "multivalued": True,
-            "options": getVocabularyTermsForForm(
-                context=portal,
-                vocab_name="rer.ufficiostampa.vocabularies.all_arguments",
-            ),
+            "options": legislatures,
+            "default": [legislatures[0]["value"]],
+            "slave": {
+                "id": "arguments",
+                "label": translate(
+                    _("legislature_arguments_label", default="Arguments"),
+                    context=request,
+                ),
+                "help": "",
+                "type": "select",
+                "multivalued": True,
+                "slaveOptions": getArguments()
+                # "options": getVocabularyTermsForForm(
+                #     context=portal,
+                #     vocab_name="rer.ufficiostampa.vocabularies.all_arguments",
+                # ),
+            },
         },
     ]